+++ lib/main.dart
@@ -2,18 +2,20 @@ import 'dart:ui';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
 import 'package:tms/general_screens/Admin/Utilities/admin_navbar.dart';
-import 'package:tms/general_screens/Admin/admin_accounts.dart';
+import 'package:tms/general_screens/Admin/Screens/Accounts/admin_accounts.dart';
+import 'package:tms/general_screens/Tutor/Add%20Lesson/tutor_lesson_request.dart';
 import 'package:tms/general_screens/Student/Utilities/student_navbar.dart';
-import 'package:tms/general_screens/Student/student_balance.dart';
+import 'package:tms/general_screens/Student/Screens/Profile%20Page/Utilities/student_balance.dart';
 import 'package:tms/general_screens/Student/this_months_learning.dart';
 import 'package:tms/general_screens/Student/welldone_page.dart';
 import 'package:tms/general_screens/Tutor/Utilities/tutor_navbar.dart';
-import 'package:tms/general_screens/Tutor/tutor_profile.dart';
-import 'package:tms/general_screens/Tutor/tutor_seerequests.dart';
-import 'package:tms/general_screens/profile%20page/subjects_page_tutor.dart';
-import 'general_screens/Admin/admin_profile.dart';
-import 'general_screens/profile page/subjects_page_student.dart';
-import 'general_screens/Admin/Utilities/create_new_acc.dart';
+import 'package:tms/general_screens/Tutor/Profile%20Page/tutor_profile.dart';
+import 'package:tms/general_screens/Tutor/Add%20Lesson/tutor_seerequests.dart';
+import 'package:tms/general_screens/Tutor/Profile%20Page/Utilities/subjects_teach.dart';
+import 'package:tms/general_screens/Tutor/Add%20Lesson/lesson_request.dart';
+import 'general_screens/Admin/Screens/Profile Page/admin_profile.dart';
+import 'general_screens/Student/Screens/Profile Page/Utilities/subjects_study.dart';
+import 'general_screens/Admin/Screens/Accounts/Utilities/create_new_acc.dart';
 
 Future main() async {
   //add both to connect to Firebase for authentication services
@@ -31,7 +33,7 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       debugShowCheckedModeBanner: false, //dont show banner in corner
-      home: TutorNavBar(),
+      home: StudentNavBar(),
     );
   }
 }
